RA8876 Controller support. I2C Encoders are broke!

+++ SDR_RA8875/SDR_RA8875.h
@@ -333,10 +333,10 @@ struct AGC {
 
 // Noise Blanker Settings
 struct NB {
-    char        nb_name[10];        // A friendly name for display
-    float       nb_threshold;       // threshold recommended to be between 1.5 and 20, closer to 3 maybe best.
-    float       nb_nAnticipation;   // nAnticipation is 1 to 125
-    float       nb_decay;           // Decay is 1 to 10.
+    char            nb_name[10];        // A friendly name for display
+    float32_t      nb_threshold;       // threshold recommended to be between 1.5 and 20, closer to 3 maybe best.
+    uint16_t        nb_nAnticipation;   // nAnticipation is 1 to 125
+    uint16_t        nb_decay;           // Decay is 1 to 10.
 };
 
 // per-band settings for common user adjustments that are band dependent. The index is the band number.
